pyproject.toml
[project]
name = "TinyExp"
version = "0.0.1"
description = "A minimalist Python project for deep learning experiment management. It uses Ray for core distributed environment and backend setup, and provides basic, no-frills tracking for models, optimizers, and LR schedulers."
authors = [{ name = "Zeming LI", email = "zane.li@connect.ust.hk" }]
readme = "README.md"
license = { file = "LICENSE" }
keywords = ['python']
requires-python = ">=3.9,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "loguru>=0.7.3",
    "numpy",
    "tabulate",
    "hydra-core",
    "omegaconf",
    "ray",
    "tqdm",
    "torch",
    "torchvision",
    "redis",
    "wandb",
    "loguru",
    "accelerate",
]

[project.urls]
Homepage = "https://zengarden.github.io/TinyExp/"
Repository = "https://github.com/zengarden/TinyExp"
Documentation = "https://zengarden.github.io/TinyExp/"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",

    "mypy>=0.991",
    "pytest-cov>=4.0.0",
    "ruff>=0.11.5",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tinyexp"]

[tool.hatch.build.targets.sdist]
include = [
    "/tinyexp",
    "/tests",
    "/CONTRIBUTING.md",
    "/LICENSE",
    "/Makefile",
    "/pyproject.toml",
    "/README.md",
]



[tool.mypy]
files = ["tinyexp"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
"tinyexp/utils/redis_utils.py" = ["S603"]

[tool.ruff.format]
preview = true

[tool.pylint.main]
max-line-length = 120
disable = "C0114, C0115, C0116"

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["tinyexp"]
omit = [
    # Example scripts are not part of the library API surface and are exercised via subprocesses.
    "tinyexp/examples/*",
    "*/tinyexp/examples/*",
    "*/working_dir_files/_ray_pkg_*/tinyexp/examples/*",
    "*/runtime_resources/working_dir_files/_ray_pkg_*/tinyexp/examples/*",
]

[tool.coverage.paths]
source = [
    # Local source tree.
    "tinyexp",
    # Installed package locations (CI / virtualenvs).
    "*/site-packages/tinyexp",
    "*/dist-packages/tinyexp",
    # Ray runtime_env packaged working directories.
    "*/working_dir_files/_ray_pkg_*/tinyexp",
    "*/runtime_resources/working_dir_files/_ray_pkg_*/tinyexp",
]

[tool.isort]
profile = "black"
line_length = 120
py_version = 310
known_first_party = "tinyexp"
known_third_party = "wandb"
